Move new code to a different file.
