+++ wfe/web-front-end_test.go
@@ -7,7 +7,6 @@ package wfe
 
 import (
 	"bytes"
-	"crypto/ecdsa"
 	"crypto/rsa"
 	"crypto/x509"
 	"encoding/json"
@@ -1386,157 +1385,6 @@ func TestVerifyPOSTUsesStoredKey(t *testing.T) {
 	// registration with test2Key
 	_, _, _, err := wfe.verifyPOST(newRequestEvent(), makePostRequest(signRequest(t, `{"resource":"foo"}`, &wfe.nonceService)), true, "foo")
 	test.AssertError(t, err, "No error returned when provided key differed from stored key.")
-}
-
-func TestRejectsNone(t *testing.T) {
-	wfe, _ := setupWFE(t)
-	_, _, _, err := wfe.verifyPOST(newRequestEvent(), makePostRequest(`
-		{
-			"header": {
-				"alg": "none",
-				"jwk": {
-					"kty": "RSA",
-					"n": "vrjT",
-					"e": "AQAB"
-				}
-			},
-			"payload": "aGkK",
-			"signature": ""
-		}
-	`), true, "foo")
-	if err == nil {
-		t.Fatalf("verifyPOST did not reject JWS with alg: 'none'")
-	}
-	if err.Error() != "algorithm 'none' in JWS header not acceptable" {
-		t.Fatalf("verifyPOST rejected JWS with alg: 'none', but for wrong reason: %s", err)
-	}
-}
-
-func TestRejectsHS256(t *testing.T) {
-	wfe, _ := setupWFE(t)
-	_, _, _, err := wfe.verifyPOST(newRequestEvent(), makePostRequest(`
-		{
-			"header": {
-				"alg": "HS256",
-				"jwk": {
-					"kty": "RSA",
-					"n": "vrjT",
-					"e": "AQAB"
-				}
-			},
-			"payload": "aGkK",
-			"signature": ""
-		}
-	`), true, "foo")
-	if err == nil {
-		t.Fatalf("verifyPOST did not reject JWS with alg: 'HS256'")
-	}
-	expected := "algorithm 'HS256' in JWS header not acceptable"
-	if err.Error() != expected {
-		t.Fatalf("verifyPOST rejected JWS with alg: 'none', but for wrong reason: got '%s', wanted %s", err, expected)
-	}
-}
-
-func TestCheckAlgorithmES(t *testing.T) {
-	stat, err := checkAlgorithm(&jose.JsonWebKey{
-		Algorithm: "ES256",
-		Key:       &ecdsa.PublicKey{},
-	}, &jose.JsonWebSignature{})
-	expected := "no signature algorithms suitable for given key type"
-	if err == nil || err.Error() != expected {
-		t.Errorf("ES256 key: Expected error '%s', got '%s'", expected, err)
-	}
-	if stat != "NoAlgorithmForKey" {
-		t.Errorf("ES256 key: Expected stat '%s', got '%s'", "NoAlgorithmForKey", stat)
-	}
-}
-
-func TestCheckAlgorithmHS(t *testing.T) {
-	stat, err := checkAlgorithm(&jose.JsonWebKey{
-		Algorithm: "HS256",
-	}, &jose.JsonWebSignature{})
-	expected := "no signature algorithms suitable for given key type"
-	if err == nil || err.Error() != expected {
-		t.Errorf("HS256 key: Expected error '%s', got '%s'", expected, err)
-	}
-	if stat != "NoAlgorithmForKey" {
-		t.Errorf("HS256 key: Expected stat '%s', got '%s'", "NoAlgorithmForKey", stat)
-	}
-}
-
-func TestCheckAlgorithmBadJWS(t *testing.T) {
-	stat, err := checkAlgorithm(&jose.JsonWebKey{
-		Key: &rsa.PublicKey{},
-	}, &jose.JsonWebSignature{
-		Signatures: []jose.Signature{
-			jose.Signature{
-				Header: jose.JoseHeader{
-					Algorithm: "HS256",
-				},
-			},
-		},
-	})
-	expected := "algorithm 'HS256' in JWS header not acceptable"
-	if err == nil || err.Error() != expected {
-		t.Errorf("HS256 JWS: Expected error '%s', got '%s'", expected, err)
-	}
-	if stat != "InvalidJWSAlgorithm" {
-		t.Errorf("HS256 JWS: Expected stat '%s', got '%s'", "InvalidJWSAlgorithm", stat)
-	}
-}
-
-func TestCheckAlgorithmBadJWK(t *testing.T) {
-	stat, err := checkAlgorithm(&jose.JsonWebKey{
-		Algorithm: "HS256",
-		Key:       &rsa.PublicKey{},
-	}, &jose.JsonWebSignature{
-		Signatures: []jose.Signature{
-			jose.Signature{
-				Header: jose.JoseHeader{
-					Algorithm: "RS256",
-				},
-			},
-		},
-	})
-	expected := "algorithm 'HS256' on JWK is unacceptable"
-	if err == nil || err.Error() != expected {
-		t.Errorf("HS256 JWS: Expected error '%s', got '%s'", expected, err)
-	}
-	if stat != "InvalidAlgorithmOnKey" {
-		t.Errorf("HS256 JWS: Expected stat '%s', got '%s'", "InvalidAlgorithmOnKey", stat)
-	}
-}
-
-func TestCheckAlgorithmSuccess(t *testing.T) {
-	_, err := checkAlgorithm(&jose.JsonWebKey{
-		Algorithm: "RS256",
-		Key:       &rsa.PublicKey{},
-	}, &jose.JsonWebSignature{
-		Signatures: []jose.Signature{
-			jose.Signature{
-				Header: jose.JoseHeader{
-					Algorithm: "RS256",
-				},
-			},
-		},
-	})
-	if err != nil {
-		t.Errorf("RS256 key: Expected nil error, got '%s'", err)
-	}
-	_, err = checkAlgorithm(&jose.JsonWebKey{
-		Key: &rsa.PublicKey{},
-	}, &jose.JsonWebSignature{
-		Signatures: []jose.Signature{
-			jose.Signature{
-				Header: jose.JoseHeader{
-					Algorithm: "RS256",
-				},
-			},
-		},
-	})
-	if err != nil {
-		t.Errorf("RS256 key: Expected nil error, got '%s'", err)
-	}
 }
 
 func TestBadKeyCSR(t *testing.T) {
